Fixes:    Compilation on amd64
[git-p4: depot-paths = "//depot/qt/main/": change = 247041]

+++ tests/auto/qscriptqobject/tst_qscriptqobject.cpp
@@ -309,7 +309,7 @@ QScriptValue fromContainer(QScriptEngine *eng, const Container &cont)
     typename Container::const_iterator end = cont.end();
     typename Container::const_iterator it;
     for (it = begin; it != end; ++it)
-        a.setProperty(eng->scriptValue(it - begin).toString(), qScriptValueFromValue(eng, *it));
+        a.setProperty(eng->scriptValue(quint32(it - begin)).toString(), qScriptValueFromValue(eng, *it));
     return a;
 }
